requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
blosc=1.19.0=h7bd577a_0
bzip2=1.0.8=he774522_0
ca-certificates=2020.6.24=0
certifi=2020.6.20=py37_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=py37_0
freetype=2.10.2=hd328e21_0
h5py=2.10.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
intel-openmp=2020.1=216
joblib=0.15.1=py_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.2.0=py37h74a9793_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=h62dcd97_0
lzo=2.10=vc14h0a64fa6_1
matplotlib=3.2.2=0
matplotlib-base=3.2.2=py37h64f37c6_0
mkl=2020.1=216
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.1.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
mock=4.0.2=py_0
ninja=1.9.0=py37h74a9793_0
numexpr=2.7.1=py37h25d0782_0
numpy=1.18.5=py37h6530119_0
numpy-base=1.18.5=py37hc3f5095_0
olefile=0.46=py37_0
openssl=1.1.1g=he774522_0
pillow=7.1.2=py37hcc1f983_0
pip=20.1.1=py37_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
pytables=3.6.1=py37h1da0976_0
python=3.7.7=h81c818b_4
python-dateutil=2.8.1=py_0
pytorch=1.5.1=py3.7_cuda102_cudnn7_0
pyyaml=5.3.1=py37he774522_1
qt=5.9.7=vc14h73c81de_0
scikit-learn=0.23.1=py37h25d0782_0
scipy=1.5.0=py37h9439919_0
setuptools=47.3.1=py37_0
sip=4.19.8=py37h6538335_0
six=1.15.0=py_0
snappy=1.1.7=vc14h2dea872_1
sqlite=3.32.3=h2a8f88b_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
torchvision=0.6.1=py37_cu102
tornado=6.0.4=py37he774522_1
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_2
wheel=0.34.2=py37_0
wincertstore=0.2=py37_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.4=ha9fde0e_3
